# docs/swagger.yaml
basePath: /api
definitions:
...
  controller.GetNextCardResponse:
    properties:
      card_id:
        type: string
    type: object
  controller.GetSessionStatsResponse:
    properties:
      card_stats:
        items:
          $ref: '#/definitions/types.CardStats'
        type: array
      current_index:
        type: integer
      remaining:
        type: integer
      total_cards:
        type: integer
      viewed_count:
        type: integer
    type: object
...
  controller.StartSessionRequest:
    properties:
      count:
        minimum: 1
        type: integer
      deck_id:
        type: string
      method:
        enum:
        - Random
        - Fails
        - Skips
        - Worst
        type: string
    required:
    - deck_id
    - method
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
  types.Card:
    properties:
      back:
        $ref: '#/definitions/types.CardBack'
      created_at:
        type: string
      deck_id:
        type: string
      fail_count:
        type: integer
      front:
        $ref: '#/definitions/types.CardFront'
      id:
        type: string
      link:
        type: string
      pass_count:
        type: integer
      retired:
        type: boolean
      reviewed_at:
        type: string
      skip_count:
        type: integer
      star_rating:
        type: integer
      updated_at:
        type: string
    type: object
  types.CardBack:
    properties:
      text:
        type: string
    type: object
  types.CardFront:
    properties:
      text:
        type: string
    type: object
  types.CardStats:
    properties:
      card_id:
        type: string
      failed:
        type: boolean
      passed:
        type: boolean
      skipped:
        type: boolean
      viewed:
        type: boolean
    type: object
  types.Deck:
    properties:
      cards:
        items:
          $ref: '#/definitions/types.Card'
        type: array
      description:
        description: New Description field
        type: string
      id:
        description: UUID string as the primary key
        type: string
      name:
        type: string
    type: object
host: 192.168.86.176:8789
info:
  contact: {}
  description: API documentation for the MeowMorize Flashcard App.
  title: MeowMorize Flashcard API
  version: "1.0"
paths:
...
  /decks/export/{id}:
    get:
      description: Export a deck as a JSON file
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Deck'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Export a deck
      tags:
      - decks
  /decks/import:
    post:
      consumes:
      - multipart/form-data
      description: Import a new deck by uploading a JSON file
      parameters:
      - description: Deck JSON File
        in: formData
        name: deck_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Deck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Import a deck from a JSON file
      tags:
      - Decks
...
  /sessions/clear:
    delete:
      description: Terminate and clear the current review session for a specific deck
      parameters:
      - description: Deck ID
        in: query
        name: deck_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clear a review session
      tags:
      - Sessions
  /sessions/next:
    get:
      description: Retrieve the ID of the next card to review in the current session
      parameters:
      - description: Deck ID
        in: query
        name: deck_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetNextCardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get the next card in the session
      tags:
      - Sessions
  /sessions/start:
    post:
      consumes:
      - application/json
      description: Initiate a new review session for a specific deck with the given
        parameters
      parameters:
      - description: Session Parameters
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/controller.StartSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a new review session
      tags:
      - Sessions
  /sessions/stats:
    get:
      description: Retrieve the statistics of the current review session for a specific
        deck
      parameters:
      - description: Deck ID
        in: query
        name: deck_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetSessionStatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get session statistics
      tags:
      - Sessions
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"